feat(core, script): rename DockerTaskRunner to Docker and ProcessTaskRunner to Process
Fixes #3729
