+++ script/src/test/java/io/kestra/plugin/scripts/runner/docker/DockerTest.java
@@ -1,0 +1,12 @@
+package io.kestra.plugin.scripts.runner.docker;
+
+import io.kestra.core.models.tasks.runners.AbstractTaskRunnerTest;
+import io.kestra.core.models.tasks.runners.TaskRunner;
+
+
+class DockerTest extends AbstractTaskRunnerTest {
+    @Override
+    protected TaskRunner taskRunner() {
+        return Docker.builder().image("centos").build();
+    }
+}
